+++ src/Adapter/Completed/CompletedElement.mjs
@@ -6,6 +6,7 @@ import { TitleElement } from "../Title/TitleElement.mjs";
 
 /** @typedef {import("../Post/backFunction.mjs").backFunction} backFunction */
 /** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
+/** @typedef {import("../../Libs/flux-localization-api/src/Adapter/Api/LocalizationApi.mjs").LocalizationApi} LocalizationApi */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
 
@@ -19,31 +20,39 @@ export class CompletedElement extends HTMLElement {
      */
     #css_api;
     /**
+     * @type {LocalizationApi}
+     */
+    #localization_api;
+    /**
      * @type {ShadowRoot}
      */
     #shadow;
 
     /**
      * @param {CssApi} css_api
+     * @param {LocalizationApi} localization_api
      * @param {backFunction | null} back_function
      * @returns {CompletedElement}
      */
-    static new(css_api, back_function = null) {
+    static new(css_api, localization_api, back_function = null) {
         return new this(
             css_api,
+            localization_api,
             back_function
         );
     }
 
     /**
      * @param {CssApi} css_api
+     * @param {LocalizationApi} localization_api
      * @param {backFunction | null} back_function
      * @private
      */
-    constructor(css_api, back_function) {
+    constructor(css_api, localization_api, back_function) {
         super();
 
         this.#css_api = css_api;
+        this.#localization_api = localization_api;
         this.#back_function = back_function;
 
         this.#shadow = this.attachShadow({ mode: "closed" });
@@ -61,16 +70,21 @@ export class CompletedElement extends HTMLElement {
     #render() {
         this.#shadow.appendChild(TitleElement.new(
             this.#css_api,
-            "Registration completed"
+            this.#localization_api.translate(
+                "Registration completed"
+            )
         ));
 
         this.#shadow.appendChild(SubtitleElement.new(
             this.#css_api,
-            "Thank you for your registration"
+            this.#localization_api.translate(
+                "Thank you for your registration"
+            )
         ));
 
         this.#shadow.appendChild(FormElement.new(
-            this.#css_api
+            this.#css_api,
+            this.#localization_api
         )
             .addButtons(
                 null,